resumimos los resultados por epochs
